Revisions to support Teensy 3.x and Teensy 4.x

+++ DCCEXParser.h
@@ -22,7 +22,7 @@
 #include "FSH.h"
 #include "RingStream.h"
 
-typedef void (*FILTER_CALLBACK)(Print * stream, byte & opcode, byte & paramCount, int p[]);
+typedef void (*FILTER_CALLBACK)(Print * stream, byte & opcode, byte & paramCount, int16_t p[]);
 typedef void (*AT_COMMAND_CALLBACK)(const byte * command);
 
 struct DCCEXParser
@@ -39,18 +39,18 @@ struct DCCEXParser
  
    private:
   
-    static const int MAX_BUFFER=50;  // longest command sent in
+    static const int16_t MAX_BUFFER=50;  // longest command sent in
      byte  bufferLength=0;
      bool  inCommandPayload=false;
      byte  buffer[MAX_BUFFER+2]; 
-    int splitValues( int result[MAX_COMMAND_PARAMS], const byte * command);
-    int splitHexValues( int result[MAX_COMMAND_PARAMS], const byte * command);
+    int16_t splitValues( int16_t result[MAX_COMMAND_PARAMS], const byte * command);
+    int16_t splitHexValues( int16_t result[MAX_COMMAND_PARAMS], const byte * command);
      
-     bool parseT(Print * stream, int params, int p[]);
-     bool parseZ(Print * stream, int params, int p[]);
-     bool parseS(Print * stream,  int params, int p[]);
-     bool parsef(Print * stream,  int params, int p[]);
-     bool parseD(Print * stream,  int params, int p[]);
+     bool parseT(Print * stream, int16_t params, int16_t p[]);
+     bool parseZ(Print * stream, int16_t params, int16_t p[]);
+     bool parseS(Print * stream,  int16_t params, int16_t p[]);
+     bool parsef(Print * stream,  int16_t params, int16_t p[]);
+     bool parseD(Print * stream,  int16_t params, int16_t p[]);
 
      static Print * getAsyncReplyStream();
      static void commitAsyncReplyStream();
@@ -60,19 +60,19 @@ struct DCCEXParser
     static Print * stashStream;
     static RingStream * stashRingStream;
     
-    static int stashP[MAX_COMMAND_PARAMS];
-    bool stashCallback(Print * stream, int p[MAX_COMMAND_PARAMS], RingStream * ringStream);
-    static void callback_W(int result);
-    static void callback_B(int result);        
-    static void callback_R(int result);
-    static void callback_Rloco(int result);
-    static void callback_Wloco(int result);
-    static void callback_Vbit(int result);
-    static void callback_Vbyte(int result);
+    static int16_t stashP[MAX_COMMAND_PARAMS];
+    bool stashCallback(Print * stream, int16_t p[MAX_COMMAND_PARAMS], RingStream * ringStream);
+    static void callback_W(int16_t result);
+    static void callback_B(int16_t result);        
+    static void callback_R(int16_t result);
+    static void callback_Rloco(int16_t result);
+    static void callback_Wloco(int16_t result);
+    static void callback_Vbit(int16_t result);
+    static void callback_Vbyte(int16_t result);
     static FILTER_CALLBACK  filterCallback;
     static FILTER_CALLBACK  filterRMFTCallback;
     static AT_COMMAND_CALLBACK  atCommandCallback;
-    static void funcmap(int cab, byte value, byte fstart, byte fstop);
+    static void funcmap(int16_t cab, byte value, byte fstart, byte fstop);
 
 };
